feat: home button component added
