+++ src/pages/components/dashboard/DashboardPresentation.jsx
@@ -1,11 +1,7 @@
-import { Link } from "react-router-dom";
-import Content from "../../../Tools/Content";
-import { FaGithub } from "react-icons/fa";
 import { Col } from "react-bootstrap";
 import { IoIosArrowDown } from "react-icons/io";
 import { motion } from "framer-motion";
-
-const key = Object.keys(Content().framework.versions).shift();
+import GetStartedButtons from "./GetStartedButtons";
 
 export default function DashboardPresentation() {
   return (
@@ -20,23 +16,7 @@ export default function DashboardPresentation() {
           single page application or a mobile application.
         </p>
 
-        <div className="d-grid gap-2 d-sm-flex justify-content-sm-center mb-5">
-          <Link
-            to={`/docs/framework/${key}/getting-started/about-as`}
-            className="btn btn-orange-custom btn-lg px-4 gap-3"
-          >
-            Get started
-          </Link>
-
-          <a
-            href="https://github.com/lion-packages/framework"
-            className="btn btn-white-custom btn-lg px-4"
-            target="_blank"
-          >
-            <FaGithub className="me-2" size={"1.3em"} />
-            Github
-          </a>
-        </div>
+        <GetStartedButtons />
 
         <motion.a
           whileHover={{ scale: 1.2, translateY: 10 }}
